Move hasIPTablesEnabled check into firewall_linux.go
